requirements.txt
Pillow
flask
flake8
